requirements.txt
streamlit
st-annotated-text
requests
python-dotenv